+++ resources/release-notes.md
@@ -1,3 +1,5 @@
 ## App
 
 - (All) Notifications handler manages warnings stack trace
+- (All) Add tooltips to `datetime` slider and `datetime` range slider
+- (All) Do not set `Limits` for `datetime` slider and `datetime` range slider if they are the same